pyproject.toml
[project]
name = "ranking"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "supabase>=2.28.0",
]
